compile warning fix.

+++ src/server/shared/Packets/ByteBuffer.h
@@ -174,7 +174,7 @@ class ByteBuffer
 
         void eraseFirst(int num)
         {
-            if (_storage.size() >= num)
+            if ((int)_storage.size() >= num)
             {
                 _storage.erase(_storage.begin(), _storage.begin() + num);
                 _storage.shrink_to_fit();
